Rearrange data member to suit Borland compiler; a few conversions unsigned int -> XMLSize_t
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@883368 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/internal/XTemplateSerializer.cpp
@@ -2168,7 +2168,7 @@ void XTemplateSerializer::storeObject(NameIdPool<DTDElementDecl>* const objToSto
     {
         NameIdPoolEnumerator<DTDElementDecl> e(objToStore, objToStore->getMemoryManager());
 
-        serEng<<e.size();
+        serEng<<(unsigned int)e.size();
 
         while (e.hasMoreElements())
         {
@@ -2221,7 +2221,7 @@ void XTemplateSerializer::storeObject(NameIdPool<DTDEntityDecl>* const objToStor
     {
         NameIdPoolEnumerator<DTDEntityDecl> e(objToStore, objToStore->getMemoryManager());
 
-        serEng<<e.size();
+        serEng<<(unsigned int)e.size();
 
         while (e.hasMoreElements())
         {
@@ -2273,7 +2273,7 @@ void XTemplateSerializer::storeObject(NameIdPool<XMLNotationDecl>* const objToSt
     {
         NameIdPoolEnumerator<XMLNotationDecl> e(objToStore, objToStore->getMemoryManager());
 
-        serEng<<e.size();
+        serEng<<(unsigned int)e.size();
 
         while (e.hasMoreElements())
         {
